implemented user auth for services and added new services as per requiremenet

+++ ProductCatalogService/src/main/java/org/example/productcatalogservice/repository/CategoryRepo.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CategoryRepo extends JpaRepository<Category, Integer> {
-    Optional<Category> findById(int id);
+public interface CategoryRepo extends JpaRepository<Category, Long> {
+    Optional<Category> findById(long id);
 
 
     Optional<Category> findByName(String name);
